+++ FirebaseDatabase/Swift/Sources/Codable/DataSnapshot+ReadDecodable.swift
@@ -15,7 +15,9 @@
  */
 
 import Foundation
-import FirebaseDatabase
+#if SWIFT_PACKAGE
+  @_exported import FirebaseDatabaseInternal
+#endif // SWIFT_PACKAGE
 import FirebaseSharedSwift
 
 public extension DataSnapshot {