+++ FirebaseDatabaseSwift/Sources/FirebaseDatabaseSwift.swift
@@ -1,0 +1,28 @@
+// Copyright 2023 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#warning("""
+The FirebaseDatabaseSwift module is deprecated and will be removed in
+the future. All of the public API from FirebaseDatabaseSwift can now be
+accessed through the FirebaseDatabase module. To migrate, delete imports
+of FirebaseDatabaseSwift and remove the module as a dependency to your
+project. If applicable, any APIs namespaced with `FirebaseDatabaseSwift`
+can now be namespaced with `FirebaseDatabase`. Additionally, if
+applicable, `@testable import FirebaseDatabaseSwift` should be replaced
+with `@testable import FirebaseDatabase`.
+""")
+
+// The `@_exported` is needed to prevent breaking clients that are using
+// types prefixed with the `FirebaseDatabase` namespace.
+@_exported import struct FirebaseDatabase.ServerTimestamp